refactor: Adjust service namespaces
