+++ SyntriceEShop.API/Controllers/IUserController.cs
@@ -1,7 +1,7 @@
 using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
 using SyntriceEShop.API.Models.UserModel;
-using SyntriceEShop.API.Services.UserServices.Models;
+using SyntriceEShop.API.Services.Models;
 
 namespace SyntriceEShop.API.Controllers;
