+++ SyntriceEShop.API/Controllers/UserController.cs
@@ -5,8 +5,8 @@ using Microsoft.Extensions.Options;
 using SyntriceEShop.API.ApplicationOptions;
 using SyntriceEShop.API.Models.UserModel;
 using SyntriceEShop.API.Services;
-using SyntriceEShop.API.Services.UserServices;
-using SyntriceEShop.API.Services.UserServices.Models;
+using SyntriceEShop.API.Services.Interfaces;
+using SyntriceEShop.API.Services.Models;
 
 namespace SyntriceEShop.API.Controllers;
