+++ SyntriceEShop.API/Services/Interfaces/IJWTProvider.cs
@@ -1,7 +1,7 @@
 using SyntriceEShop.API.Models.RefreshTokenModel;
 using SyntriceEShop.API.Models.UserModel;
 
-namespace SyntriceEShop.API.Services.UserServices;
+namespace SyntriceEShop.API.Services.Interfaces;
 
 public interface IJWTProvider
 {
