+++ SyntriceEShop.Tests/API/Services/UserServices/UserServiceTests.cs
@@ -4,9 +4,11 @@ using Shouldly;
 using SyntriceEShop.API.Models.RefreshTokenModel;
 using SyntriceEShop.API.Models.UserModel;
 using SyntriceEShop.API.Repositories;
+using SyntriceEShop.API.Repositories.Interfaces;
 using SyntriceEShop.API.Services;
-using SyntriceEShop.API.Services.UserServices;
-using SyntriceEShop.API.Services.UserServices.Models;
+using SyntriceEShop.API.Services.Implementations;
+using SyntriceEShop.API.Services.Interfaces;
+using SyntriceEShop.API.Services.Models;
 
 namespace SyntriceEShop.Tests.API.Services.UserServices;
